- Added flag -msse2 for SSE instructions in C++ - Warning fix in DRAM
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@2901 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/dram/Channel.cc
@@ -102,7 +102,7 @@ void Channel::Scheduler()
 void Channel::dump(std::ostream &os) const
 {
 	os << misc::fmt("\tDumping Channel %d\n", id);
-	os << misc::fmt("\t%d Ranks\n\tRank dump:\n", ranks.size());
+	os << misc::fmt("\t%d Ranks\n\tRank dump:\n", (int) ranks.size());
 
 	for (auto rank : ranks)
 	{
